feature: create endpoints to create project and demands

+++ demand-capacity-mgmt-backend/src/main/java/org/eclipse/tractusx/demandcapacitymgmt/demandcapacitymgmtbackend/repositories/ProjectRepository.java
@@ -5,6 +5,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProjectRepository extends JpaRepository<ProjectEntity, Integer> {}
-
-
+public interface ProjectRepository extends JpaRepository<ProjectEntity, Long> {}
